Fixups
- Fixed some problems in the DOF example
- Fixed a problem in the Cube shape builder

+++ include/oglplus/shapes/cube.hpp
@@ -390,7 +390,7 @@ public:
 				{
 					DrawOperation::Method::DrawElements,
 					PrimitiveType::LineLoop,
-					r*4, 4
+					GLuint(r*4), GLuint(4)
 				}
 			);
 		}
